+++ MapboxSearch/discover/src/test/java/com/mapbox/search/discover/DiscoverImplTest.kt
@@ -3,17 +3,24 @@ package com.mapbox.search.discover
 import com.mapbox.bindgen.ExpectedFactory
 import com.mapbox.geojson.BoundingBox
 import com.mapbox.geojson.Point
+import com.mapbox.search.base.BaseSearchCallback
 import com.mapbox.search.base.core.createCoreSearchOptions
 import com.mapbox.search.base.result.BaseSearchResult
 import com.mapbox.search.base.result.BaseSearchResultType
 import com.mapbox.search.base.result.BaseServerSearchResultImpl
 import com.mapbox.search.base.result.mapToBase
+import com.mapbox.search.base.task.AsyncOperationTaskImpl
 import com.mapbox.search.base.utils.extension.mapToCore
+import com.mapbox.search.common.CompletionCallback
 import com.mapbox.search.common.IsoLanguageCode
+import com.mapbox.search.common.TestExecutor
 import com.mapbox.search.common.createTestCoreSearchResult
 import io.mockk.coEvery
 import io.mockk.coVerify
+import io.mockk.every
 import io.mockk.mockk
+import io.mockk.slot
+import io.mockk.verify
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertSame
@@ -27,6 +34,8 @@ internal class DiscoverImplTest {
 
     private lateinit var discover: Discover
     private lateinit var engine: DiscoverSearchEngine
+
+    private val testExecutor = TestExecutor()
 
     @BeforeEach
     fun setUp() {
@@ -81,6 +90,35 @@ internal class DiscoverImplTest {
     }
 
     @Test
+    fun `Check request parameters for callback-based nearby search`() {
+        val slotCallback = slot<BaseSearchCallback>()
+        every { engine.search(any(), any(), any(), capture(slotCallback)) } answers {
+            slotCallback.captured.onError(Exception())
+            mockk()
+        }
+
+        val defaultOptions = DiscoverOptions()
+
+        val callback = mockk<CompletionCallback<List<DiscoverResult>>>(relaxed = true)
+        discover.search(TEST_QUERY, TEST_USER_LOCATION, defaultOptions, testExecutor, callback)
+
+        val coreOptions = createCoreSearchOptions(
+            proximity = TEST_USER_LOCATION,
+            limit = defaultOptions.limit,
+            language = listOf(defaultOptions.language.code)
+        )
+
+        verify(exactly = 1) {
+            engine.search(
+                eq(TEST_QUERY.canonicalName),
+                eq(coreOptions),
+                eq(testExecutor),
+                eq(slotCallback.captured)
+            )
+        }
+    }
+
+    @Test
     fun `Check request parameters for search in region`() {
         coEvery { engine.search(any(), any()) } returns ExpectedFactory.createError(Exception())
 
@@ -124,6 +162,36 @@ internal class DiscoverImplTest {
             engine.search(
                 eq(TEST_QUERY.canonicalName),
                 eq(coreOptions)
+            )
+        }
+    }
+
+    @Test
+    fun `Check request parameters for callback-based search in region`() {
+        val slotCallback = slot<BaseSearchCallback>()
+        every { engine.search(any(), any(), any(), capture(slotCallback)) } answers {
+            slotCallback.captured.onError(Exception())
+            mockk()
+        }
+
+        val defaultOptions = DiscoverOptions()
+
+        val callback = mockk<CompletionCallback<List<DiscoverResult>>>(relaxed = true)
+        discover.search(TEST_QUERY, TEST_BOUNDING_BOX, TEST_USER_LOCATION, defaultOptions, testExecutor, callback)
+
+        val coreOptions = createCoreSearchOptions(
+            proximity = TEST_USER_LOCATION,
+            bbox = TEST_BOUNDING_BOX.mapToCore(),
+            limit = defaultOptions.limit,
+            language = listOf(defaultOptions.language.code)
+        )
+
+        verify(exactly = 1) {
+            engine.search(
+                eq(TEST_QUERY.canonicalName),
+                eq(coreOptions),
+                eq(testExecutor),
+                eq(slotCallback.captured)
             )
         }
     }
@@ -179,6 +247,37 @@ internal class DiscoverImplTest {
     }
 
     @Test
+    fun `Check request parameters for callback-based search along the route`() {
+        val slotCallback = slot<BaseSearchCallback>()
+        every { engine.search(any(), any(), any(), capture(slotCallback)) } answers {
+            slotCallback.captured.onError(Exception())
+            mockk()
+        }
+
+        val defaultOptions = DiscoverOptions()
+
+        val callback = mockk<CompletionCallback<List<DiscoverResult>>>(relaxed = true)
+        discover.search(TEST_QUERY, TEST_ROUTE, RouteDeviationOptions.DEFAULT_DEVIATION, defaultOptions, testExecutor, callback)
+
+        val coreOptions = createCoreSearchOptions(
+            limit = defaultOptions.limit,
+            language = listOf(defaultOptions.language.code),
+            route = TEST_ROUTE,
+            sarType = RouteDeviationOptions.SarType.ISOCHROME.rawName,
+            timeDeviation = RouteDeviationOptions.Time(value = 10L, unit = TimeUnit.MINUTES).timeDeviationMinutes,
+        )
+
+        verify(exactly = 1) {
+            engine.search(
+                eq(TEST_QUERY.canonicalName),
+                eq(coreOptions),
+                eq(testExecutor),
+                eq(slotCallback.captured)
+            )
+        }
+    }
+
+    @Test
     fun `Check successful search response mapping to Discover format`() {
         val coreResults = listOf(
             createTestSearchResult(point(1, 2)),
@@ -200,6 +299,30 @@ internal class DiscoverImplTest {
     }
 
     @Test
+    fun `Check callback-based successful search response mapping to Discover format`() {
+        val coreResults = listOf(
+            createTestSearchResult(point(1, 2)),
+            createTestSearchResult(point(3, 4))
+        )
+
+        val mockedTask = AsyncOperationTaskImpl<Unit>()
+        val slotCallback = slot<BaseSearchCallback>()
+        every { engine.search(any(), any(), any(), capture(slotCallback)) } answers {
+            slotCallback.captured.onResults(coreResults, mockk())
+            mockedTask
+        }
+
+        val callback = mockk<CompletionCallback<List<DiscoverResult>>>(relaxed = true)
+        val searchTask = discover.search(TEST_QUERY, TEST_USER_LOCATION, TEST_OPTIONS, testExecutor, callback)
+
+        verify(exactly = 1) {
+            callback.onComplete(eq(coreResults.map { DiscoverResult.createFromSearchResult(it) }))
+        }
+
+        assertSame(mockedTask, searchTask)
+    }
+
+    @Test
     fun `Check error search response mapping to Discover format`() {
         val error = IOException("Test Unknown Host Error")
         coEvery { engine.search(any(), any()) } returns ExpectedFactory.createError(error)
@@ -210,6 +333,27 @@ internal class DiscoverImplTest {
 
         assertTrue(result.isError)
         assertSame(error, result.error)
+    }
+
+    @Test
+    fun `Check callback-based error search response mapping to Discover format`() {
+        val error = IOException("Test Unknown Host Error")
+
+        val mockedTask = AsyncOperationTaskImpl<Unit>()
+        val slotCallback = slot<BaseSearchCallback>()
+        every { engine.search(any(), any(), any(), capture(slotCallback)) } answers {
+            slotCallback.captured.onError(error)
+            mockedTask
+        }
+
+        val callback = mockk<CompletionCallback<List<DiscoverResult>>>(relaxed = true)
+        val searchTask = discover.search(TEST_QUERY, TEST_USER_LOCATION, TEST_OPTIONS, testExecutor, callback)
+
+        verify(exactly = 1) {
+            callback.onError(eq(error))
+        }
+
+        assertSame(mockedTask, searchTask)
     }
 
     private companion object {