+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/SearchEngineFactory.kt
@@ -2,6 +2,7 @@ package com.mapbox.search
 
 import com.mapbox.search.analytics.AnalyticsService
 import com.mapbox.search.base.BaseSearchSdkInitializer
+import com.mapbox.search.base.StubCompletionCallback
 import com.mapbox.search.base.core.CoreEngineOptions
 import com.mapbox.search.base.core.CoreLocationProvider
 import com.mapbox.search.base.core.CoreSearchEngine
@@ -10,6 +11,7 @@ import com.mapbox.search.base.location.LocationEngineAdapter
 import com.mapbox.search.base.location.WrapperLocationProvider
 import com.mapbox.search.base.utils.UserAgentProvider
 import com.mapbox.search.base.utils.extension.mapToCore
+import com.mapbox.search.common.CompletionCallback
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
 import com.mapbox.search.utils.CompoundCompletionCallback
 import java.util.concurrent.Executor