+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ### New features
 - [CORE] Now `SearchSuggestion` and `SearchResult` types provide a new field `fullAddress` which contains formatted text address from backend.
+- [DISCOVER] Now `Discover` SDK provides callback-based alternatives for Kotlin suspend functions.
+- [DISCOVER] Now `DiscoverResult` provides a new field `formattedAddress`.
+
+### Breaking changes
+- [CORE] `com.mapbox.search.CompletionCallback` has been moved to `com.mapbox.search.common.CompletionCallback`.
 
 ### Bug fixes
 - [CORE] Fixed `SearchSuggestion.name` and `SearchResult.name` formatting for address types.